requirements.txt
asttokens==3.0.0
attrs==25.4.0
av==13.1.0
beautifulsoup4==4.14.2
bleach==6.3.0
cattrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.1
cloup==3.0.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
fastf1==3.6.1
fastjsonschema==2.21.2
fonttools==4.59.0
glcontext==3.0.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isosurfaces==0.1.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kagglehub==0.3.12
kiwisolver==1.4.8
manim==0.19.1
ManimPango==0.6.1
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
moderngl==5.12.0
moderngl-window==3.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycairo==1.29.0
pydub==0.25.1
pyglet==2.1.11
pyglm==2.8.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
RapidFuzz==3.14.1
referencing==0.37.0
requests==2.32.4
requests-cache==1.2.1
rich==14.2.0
rpds-py==0.28.0
scipy==1.16.3
screeninfo==0.8.1
seaborn==0.13.2
six==1.17.0
skia-pathops==0.9.0
soupsieve==2.8
srt==3.5.3
stack-data==0.6.3
svgelements==1.9.6
timple==0.1.8
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
